Develop

+++ src/modules/workshop/UI/Panel/BottomBar/index.tsx
@@ -2,30 +2,39 @@
 import { IconClear, IconRedo, IconTrash, IconUndo } from '@/components/IconSvgs'
 import { useUndoRedoShortcut } from '@/hooks/useShortcuts'
 import IcOpen from '@/icons/workshop/ic-open.svg'
-import { default as IcCreate, default as IcSave } from '@/icons/workshop/ic-save.svg'
-import { useProjectStore, useStoreGlobal } from '@/stores/blocks'
+import { default as IcSave } from '@/icons/workshop/ic-save.svg'
+import { useModalStore, useProjectStore, useStoreGlobal } from '@/stores/blocks'
 import s from './styles.module.scss'
+import IcCreate from '@/icons/workshop/ic-create.svg'
 
 import useApiInfinite from '@/hooks/useApiInfinite'
-import { getListModularByWallet } from '@/services/api/generative'
+import { getListModularByWallet, getProjectDetail, handleGetData, uploadFile } from '@/services/api/generative'
 import { TBlockData, TListCurrent } from '@/types'
 import instance from '@/utils/storage/local-storage'
 import { useAppSelector } from '@/stores/hooks'
 import { accountSelector } from '@/stores/states/wallet/selector'
 import { handleConvertData } from '@/utils/convertTraits'
-import { useEffect } from 'react'
+import { useEffect, useMemo, useRef } from 'react'
+import IcEye from '@/icons/workshop/ic-eye.svg'
 
 type TDataFetch = {
   list: TListCurrent
 }
 
-import SavedProjectsModal from '@/modules/workshop/components/Modal/SavedProjectsModal'
+import SavedProjectsModal, { SAVED_PROJECTS_MODAL_ID } from '@/modules/workshop/components/Modal/SavedProjectsModal'
 import { useId, useState } from 'react'
 import { Toaster } from 'react-hot-toast'
 import jsonFile from './mock.json'
 import UnsaveWarningModal from '@/modules/workshop/components/Modal/UnsaveWarningModal'
+import SetProjectNameModal, { SET_PROJECT_NAME_MODAL_ID } from '@/modules/workshop/components/Modal/SetProjectNameModal'
+import { SHA256 } from 'crypto-js'
+import { convertBase64ToFile } from '@/utils/file'
+import { WORKSHOP_URL } from '@/constant/route-path'
+import { useRouter } from 'next/navigation'
+import { DELAY_SNAPSHOT } from '@/constant/constant'
+import { EDIT_MODE } from '@/utils'
 
-const MOCK_ADDRESS = 'bc1p4psqwcglffqz87kl0ynzx26dtxvu3ep75a02d09fshy90awnpewqvkt7er'
+// const MOCK_ADDRESS = 'bc1p4psqwcglffqz87kl0ynzx26dtxvu3ep75a02d09fshy90awnpewqvkt7er'
 
 export default function BottomBar() {
   const {
@@ -35,20 +44,35 @@ export default function BottomBar() {
     viewPreview,
     setViewPreview,
     deleteAlls,
-    blockCurrent,
     setDataCurrent,
     setBlockCurrent,
     blocksState,
+    selectedBricks,
     // deleteSeletBlocks,
+    deleteSelected,
+    setSelectedBricks,
+    setBricks,
+    blockCurrent,
+    setBlockCurrentUpdate,
   } = useStoreGlobal()
 
-  const { projectId, saveProject, createProject, projectName, renderFile } = useProjectStore()
+  const router = useRouter()
+
+  const { setLoading, projectId, saveProject, createProject, projectName, renderFile, loadProject } = useProjectStore()
+
+
+
+  const { openModal, modals } = useModalStore()
 
   const [showModal, setShowModal] = useState(false)
   const [showUnsaveModal, setShowUnsaveModal] = useState(false)
+  const [showSetProjectNameModal, setShowSetProjectNameModal] = useState(false)
+  const [clickView, setClickView] = useState(false)
+
+  const currentBlockStateRef = useRef(SHA256(JSON.stringify(blockCurrent)).toString() || '')
 
   const account = useAppSelector(accountSelector)
-
+  const isEditMode = mode === EDIT_MODE
   const id = useId()
 
   const {
@@ -58,7 +82,8 @@ export default function BottomBar() {
   } = useApiInfinite(
     getListModularByWallet,
     {
-      ownerAddress: MOCK_ADDRESS, //account?.address,
+      ownerAddress: account?.address,
+      // ownerAddress: 'bc1pafhpvjgj5x7era4cv55zdhpl57qvj0c60z084zsl7cwlmn3gq9tq3hqdmn',
       page: 1,
       limit: 20,
     },
@@ -76,31 +101,85 @@ export default function BottomBar() {
   const redoAction = () => {
     redo()
   }
-  const deleteAction = () => {
-    // deleteSeletBlocks()
-  }
+
+  const isAllowSave = useMemo(() => {
+    const hashBlockCurrent = SHA256(JSON.stringify(blockCurrent)).toString()
+
+    return hashBlockCurrent !== currentBlockStateRef.current
+  }, [blockCurrent])
+
   const saveAction = async () => {
-    if (blocksState.length < 2 || !blockCurrent || blockCurrent.length === 0) return
+    // saveToPng()
 
-    const payload: {
-      jsonFile: any
-      projectId?: string
-      projectName?: string
-      ownerAddress: string
-    } = {
-      jsonFile: blockCurrent,
-      ownerAddress: MOCK_ADDRESS, //account?.address,
+    if (!isAllowSave) return
+
+    if (!projectName) {
+      openModal({
+        id: SET_PROJECT_NAME_MODAL_ID,
+        component: <SetProjectNameModal type='save' />,
+      })
+      return
     }
+    setLoading(true)
 
-    if (projectId) {
-      payload.projectId = projectId
-    }
+    const wrapperDom = document.querySelector('.styles_workshop_preview__cFkSM') // TODO: Pass ref to
+    // if (e.ctrlKey && e.key === 's') {
+    ;(wrapperDom as HTMLElement).style.display = 'block'
+    ;(wrapperDom as HTMLElement).style.position = 'fixed'
+    ;(wrapperDom as HTMLElement).style.top = '0'
+    ;(wrapperDom as HTMLElement).style.left = '0'
+    ;(wrapperDom as HTMLElement).style.right = '0'
+    ;(wrapperDom as HTMLElement).style.bottom = '0'
 
-    if (projectName) {
-      payload.projectName = projectName
-    }
+    const canvas = wrapperDom.querySelector('canvas')
+    canvas.classList.add(s.saveMove)
 
-    saveProject(payload)
+    setTimeout(async () => {
+      const image = canvas.toDataURL('image/png')
+      const file = convertBase64ToFile(image)
+      const resUrl = await uploadFile({ file })
+      const a = document.createElement('a')
+      a.href = image
+      a.download = 'project-xxxx.png'
+      // a.click()
+      a.remove()
+
+      canvas.classList.remove(s.saveMove)
+      ;(wrapperDom as HTMLElement).style.display = 'none'
+
+      const payload: {
+        jsonFile: any
+        projectId?: string
+        projectName?: string
+        ownerAddress: string
+        thumbnail: string
+      } = {
+        jsonFile: blockCurrent,
+        ownerAddress: account?.address,
+        thumbnail: resUrl.url,
+      }
+
+      if (projectId) {
+        payload.projectId = projectId
+      }
+
+      if (projectName) {
+        payload.projectName = projectName
+      }
+      await saveProject(payload)
+
+      setLoading(false)
+    }, DELAY_SNAPSHOT)
+  }
+
+  const saveAsAction = async () => {
+    if (!isAllowSave) return
+
+    openModal({
+      id: SET_PROJECT_NAME_MODAL_ID,
+      component: <SetProjectNameModal type='save-as' />,
+    })
+    return
   }
 
   const loadDataAction = (file) => {
@@ -115,27 +194,69 @@ export default function BottomBar() {
     deleteAlls()
   }
 
-  const handleGetData = async () => {
-    const data = (await getListModularByWallet({
-      ownerAddress: 'bc1pafhpvjgj5x7era4cv55zdhpl57qvj0c60z084zsl7cwlmn3gq9tq3hqdmn',
-      page: 1,
-      limit: 20,
-    })) as TDataFetch
-    const convertData =
-      Array.isArray(data.list) &&
-      data.list.map((item) => {
-        const traits = handleConvertData(item.attributes)
-        return { ...item, traits }
-      })
-    setDataCurrent(convertData)
-  }
+  // const handleGetData = async () => {
+  //   const data = (await getListModularByWallet({
+  //     ownerAddress: account?.address,
+  //     // ownerAddress: 'bc1pafhpvjgj5x7era4cv55zdhpl57qvj0c60z084zsl7cwlmn3gq9tq3hqdmn',
+  //     page: 1,
+  //     limit: 20,
+  //   })) as any
+  //   const listData = data.list as TListCurrent[]
+  //   return listData
+  // }
 
-  const handleClickCreateNewProject = () => {
-    // if (blocksState.length > 2) {
-    //   setShowUnsaveModal(true)
-    // }
+  const handleClickCreateNewProject = async () => {
+    if (isAllowSave) {
+      setShowUnsaveModal(true)
+      return
+    }
     createProject()
     deleteAlls()
+    const data = await handleGetData(account?.address) //reset new project
+    setDataCurrent(data)
+  }
+
+  const viewAction = async () => {
+    if (!projectId && !isAllowSave) return
+
+    if (!projectId && isAllowSave) {
+      openModal({
+        id: SET_PROJECT_NAME_MODAL_ID,
+        component: <SetProjectNameModal type='save-view' />,
+      })
+      return
+    }
+
+    if (isAllowSave) {
+      await saveAction()
+      setLoading(true)
+      setTimeout(() => {
+        router.push(`${WORKSHOP_URL}/${projectId}`)
+      }, 3000)
+      return
+    }
+    router.push(`${WORKSHOP_URL}/${projectId}`)
+  }
+
+  const loadInitialProject = async () => {
+    try {
+      setLoading(true)
+      const res = await getProjectDetail({ id: projectId })
+      // const data = await handleGetData(account.address) // reset data when open new data
+
+      if (!!res.metaData) {
+        // setDataCurrent(data)
+        loadProject({
+          projectId: projectId,
+          projectName: projectName,
+          renderFile: res.metaData,
+        })
+      }
+    } catch (error) {
+      //
+    } finally {
+      setLoading(false)
+    }
   }
 
   useUndoRedoShortcut(undo, redo)
@@ -147,41 +268,71 @@ export default function BottomBar() {
   }, [renderFile])
 
   useEffect(() => {
-    handleGetData()
-  }, [])
+    // detect click browser back button or closing tab
+    if (!isAllowSave) return
+    window.addEventListener('beforeunload', (e) => {
+      e.preventDefault()
+    })
 
-  // First load, trigger create new project
-  useEffect(() => {
-    if (!projectId) {
-      createProject()
+    return () => {
+      window.removeEventListener('beforeunload', (e) => {
+        e.preventDefault()
+      })
     }
-  }, [projectId, createProject])
+  }, [isAllowSave])
+
+  useEffect(() => {
+    if (projectId && projectName) {
+      loadInitialProject()
+    }
+  }, [projectId, projectName])
+
+  // useEffect(() => {
+
+  //   window.addEventListener('keydown', saveToPng)
+  //   return () => {
+  //     window.removeEventListener('keydown', saveToPng)
+  //   }
+  // }, [])
+
+  const handleDeleteSelected = () => {
+    deleteSelected(selectedBricks)
+    setSelectedBricks({})
+  }
 
   return (
     <>
       <Toaster />
       <div className={s.wrapper}>
         <div className={s.bottomBar}>
-          <button className={s.bottomBar_btn} onClick={undoAction}>
-            <IconUndo /> Undo
+          <button className={`${s.bottomBar_btn} ${s.noFill}`} onClick={undoAction}>
+            <IconUndo />
           </button>
-          <button className={s.bottomBar_btn} onClick={redoAction}>
-            <IconRedo /> Redo
+          <button className={`${s.bottomBar_btn} ${s.noFill}`} onClick={redoAction}>
+            <IconRedo />
           </button>
           <button className={s.bottomBar_btn} onClick={() => handleDeleteAll()}>
             <IconClear />
-            Clear
           </button>
-          <button className={s.bottomBar_btn} onClick={deleteAction}>
-            <IconTrash /> Delete
-          </button>
+          {isEditMode && (
+            <button
+              className={`${s.bottomBar_btn} ${selectedBricks.length === 0 && s.disable}`}
+              disabled={selectedBricks.length === 0 && true}
+              onClick={handleDeleteSelected}
+            >
+              <IconTrash />
+            </button>
+          )}
         </div>
 
         <div className={s.bottomBar}>
-          <button className={s.bottomBar_btn} onClick={saveAction}>
-            <IcSave /> Save Project
+          <button className={s.bottomBar_btn} onClick={viewAction}>
+            <IcEye /> View Mode
           </button>
-          <button className={s.bottomBar_btn} onClick={saveAction}>
+          <button className={s.bottomBar_btn} onClick={saveAction} disabled={!isAllowSave}>
+            <IcSave /> Save
+          </button>
+          <button className={s.bottomBar_btn} onClick={saveAsAction} disabled={!isAllowSave}>
             <IcSave /> Save As
           </button>
           <button className={s.bottomBar_btn} onClick={handleClickCreateNewProject}>
@@ -190,14 +341,16 @@ export default function BottomBar() {
           <button
             className={s.bottomBar_btn}
             onClick={() => {
-              setShowModal(true)
+              openModal({
+                id: SAVED_PROJECTS_MODAL_ID,
+                component: <SavedProjectsModal />,
+              })
             }}
           >
-            <IcOpen /> Open Project
+            <IcOpen /> Open
           </button>
         </div>
       </div>
-      <SavedProjectsModal show={showModal} setIsOpen={setShowModal} />
       <UnsaveWarningModal show={showUnsaveModal} setIsOpen={setShowUnsaveModal} />
     </>
   )
